requirements.txt
# YOLOv11x Backend Requirements
# Generated with pip freeze

# Core YOLO and Vision
ultralytics==8.3.204
opencv-python==4.12.0.88
torch==2.8.0
torchvision==0.23.0

# Server and WebSocket
python-socketio==5.14.1
python-engineio==4.12.3
websocket-client==1.8.0
uvicorn==0.37.0

# Data Processing
numpy==2.2.6
pillow==11.3.0


